+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Implemented log_level's
 - Removed methods ok, err, warn, throw_err
 - Added new method 'set_log_level_redirect_to_global' and implemented it
+- Removed set_enable_log_file and enable_log_file (can control it via set_log_level_file_output)
 
 ## [v2.0.0] - 2026-06-16 (2026 June 16)